crcCheck
This program perform error checking with the use of polynomial codes.It
takes as input a txt file.Every word of the file becomes an input for a
crc code generating function and will produce a codeword in binary.Code
words could be printed in a txt file.The user could introduce errors in
the results. The user then could check its file for errors.